AGDEV-11306 Refactor RasterBandMixins::Extensions#read_by_block
The method name implied that it read data by blocks, then yielded that
data. That was not the case: it read by blocks, but then continued to
read by row within the block and yield the row of pixels instead of the
full block full of pixels. Now we have #read_blocks_by_blocks which
does the latter, and which #read_lines_by_block uses to extract lines
from blocks and do what it did before.