AGDEV-12216 Allow points param for Gridder
Previously you could only pass in a Layer, where we’d extract all
points from that layer and use those for gridding. If, for some reason,
you want to use a different set of points (say, a subset of points),
you can now pass that set in to use for gridding.

Also, I extracted out methods that dealt with building the points array
(in the case you _don’t_ pass them in) into PointExtracting.

Lastly, I switched from using an NArray for passing points around to
just use a vanilla Array. Didn’t seem like any performance was gained
form the NArray in this case, plus made it extra work for dealing with
the NArray.