Update multi_polygon_spec.rb
This was recently changed to work with the version of GDAL that's in the gdal2 docker image. After switching the GH action back to use native ubuntu 18.04, the spec is back to failing in that action (but still passes in the docker container). There's either a difference in the version of GDAL or GDAL was built with different options, resulting in this test failure. Either way, the actual value of the test isn't important.